catadelivery/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-unfold==0.27.0
django-import-export==3.3.8
drf-spectacular==0.27.2
scipy==1.11.4
numpy==1.26.4
django-cors-headers==4.3.1
Pillow==10.4.0
firebase-admin==6.4.0
channels==4.1.0
daphne==4.1.2
pymongo==4.7.3
